requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aom=3.7.0=h463b476_0
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brotli-python=1.1.0=py311ha891d26_1
bzip2=1.0.8=h93a5062_5
ca-certificates=2023.11.17=hf0a4a13_0
cairo=1.18.0=hd1e100b_0
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=py311hd03642b_0
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hb547adb_0
debugpy=1.8.0=py311ha891d26_1
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=hb7217d7_1
ffmpeg=6.1.0=gpl_hb975139_101
filelock=3.13.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.45.0=py311h05b510d_0
freetype=2.12.1=hadb7bae_2
fribidi=1.0.10=h27ca646_0
gettext=0.21.1=h0186832_0
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.40=pyhd8ed1ab_0
gmp=6.3.0=h965bd2d_0
gmpy2=2.1.2=py311h2ba9262_1
gnutls=3.7.8=h9f1a10d_0
graphite2=1.3.13=h9f76cd9_1001
harfbuzz=8.3.0=h8f0ba13_0
icu=73.2=hc8870d7_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.26.0=pyh3cd1d5f_0
ipython=8.17.2=pyh31c8845_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.2=pyhd8ed1ab_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py311h267d04e_0
kiwisolver=1.4.5=py311he4fd1f5_1
lame=3.100=h1a8c8d9_1003
lcms2=2.15=hf2736f0_3
lerc=4.0.0=h9a09cb3_0
libabseil=20230802.1=cxx17_h13dd4ca_0
libass=0.17.1=hf7da4fe_1
libblas=3.9.0=19_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=19_osxarm64_openblas
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran5=13.2.0=hf226fd6_1
libglib=2.78.1=hb438215_1
libiconv=1.17=he4db4b2_0
libidn2=2.3.4=h1a8c8d9_0
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=19_osxarm64_openblas
libopenblas=0.3.24=openmp_hd76b1f2_0
libopus=1.3.1=h27ca646_1
libpng=1.6.39=h76d750c_0
libprotobuf=4.24.4=hc9861d8_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.44.0=h091b4b1_0
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.6.0=ha8a6c65_2
libunistring=0.9.10=h3422bc3_0
libvpx=1.13.1=hb765f3a_0
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libxml2=2.11.6=h0d0cfa8_0
libzlib=1.2.13=h53f4e23_5
lightning-utilities=0.10.0=pyhd8ed1ab_0
llvm-openmp=17.0.5=hcd81f8e_0
markupsafe=2.1.3=py311heffc1b2_1
matplotlib=3.8.2=py311ha1ab1f8_0
matplotlib-base=3.8.2=py311hfdba5f6_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h9546428_0
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h463b476_2
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.8.1=h63371fa_1
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.0=py311hb8f3215_0
openh264=2.3.1=hb7217d7_2
openjpeg=2.5.0=h4c1507b_3
openssl=3.1.4=h0d3ecfb_0
p11-kit=0.24.1=h29577a5_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.3=py311h6e08293_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=py_1
patsy=0.5.3=pyhd8ed1ab_0
pcre2=10.42=h26f9a81_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=py311hb9c5795_0
pip=23.3.1=pyhd8ed1ab_0
pixman=0.42.2=h13dd4ca_0
platformdirs=4.0.0=pyhd8ed1ab_0
prompt-toolkit=3.0.41=pyha770c72_0
prompt_toolkit=3.0.41=hd8ed1ab_0
protobuf=4.24.4=py311h4d1eceb_0
psutil=5.9.5=py311heffc1b2_1
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.17.1=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.6=h47c9636_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=4_cp311
pytorch=2.1.1=py3.11_0
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0.1=py311heffc1b2_1
pyzmq=25.1.1=py311he9c0408_2
readline=8.2=h92ec313_1
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.3.2=py311ha25ca4d_1
scipy=1.11.3=py311h93d07a4_1
seaborn=0.13.0=hd8ed1ab_0
seaborn-base=0.13.0=pyhd8ed1ab_0
sentry-sdk=1.35.0=pyhd8ed1ab_0
setproctitle=1.3.3=py311heffc1b2_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py311h9ea6feb_2
svt-av1=1.7.0=hb765f3a_0
sympy=1.12=pypyh9d50eac_103
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.13=h5083fa2_1
torchmetrics=1.2.0=pyhd8ed1ab_0
torchvision=0.16.1=py311_cpu
tornado=6.3.3=py311heffc1b2_1
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.13.0=pyhd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.1.0=pyhd8ed1ab_0
wandb=0.15.12=pyhd8ed1ab_0
wcwidth=0.2.10=pyhd8ed1ab_0
wheel=0.41.3=pyhd8ed1ab_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=h965bd2d_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=h4f39d0f_0
